requirements.txt
pyrogram==2.0.0
tgcrypto==1.2.0
python-dotenv==0.21.0
